+++ httpie/utils.py
@@ -111,3 +111,5 @@ def get_expired_cookies(
         for cookie in cookies
         if cookie.get('expires', float('Inf')) <= now
     ]
+
+